feat: Add demo.html file and hook up button for WKWebViewKeyboardFocus demo
This adds a demo.html file that contains a text field and a button that focusses the
text field when tapped. It also hooks up a native button and executes the same JS
in the webview, which does not result in the textfield being focussed or the keyboard
being shown.